refactor: XIP0028 - Extract recording and upload methods to partial classes
Reduces WorkerTask.cs from 1,171 to 703 lines (40% reduction) by moving:
- Recording methods (327 lines) to WorkerTaskRecording.cs
- Upload/clipboard methods (222 lines) to WorkerTaskUpload.cs

This improves maintainability by organizing code by domain while
maintaining the unified TaskManager → WorkerTask execution pipeline.
